[openmm] 7.4.0 beta with CUDA 10.0 as default beta channel

+++ devtools/osx-build.sh
@@ -1,9 +1,19 @@
+# upgrade Bash to version 4, for associative array support
+if [[ ${BASH_VERSINFO[0]} < 4 ]]; then
+    brew install bash
+    /usr/local/bin/bash -uxe $0
+    exit $?
+fi
+
 #!/bin/bash
 set -e -x
 export MACOSX_DEPLOYMENT_TARGET="10.9"
 # Clear existing locks
 rm -rf /usr/local/var/homebrew/locks
 # Update homebrew cant disable this yet, -y and --quiet do nothing
+brew update
+# Upgrade brew
+brew upgrade
 brew update
 
 # Install Miniconda
@@ -34,22 +44,26 @@ if [ "$INSTALL_OPENMM_PREREQUISITES" = true ] ; then
     # Install OpenMM dependencies that can't be installed through
     # conda package manager (doxygen + CUDA)
     brew install https://raw.githubusercontent.com/Homebrew/homebrew-core/5b680fb58fedfb00cd07a7f69f5a621bb9240f3b/Formula/doxygen.rb
-    # Make the nvidia-cache if not there
-    mkdir -p $NVIDIA_CACHE
-    cd $NVIDIA_CACHE
-    # Download missing nvidia installers
-    if ! [ -f cuda_mac_installer_tk.tar.gz ]; then
-        curl -O -# http://developer.download.nvidia.com/compute/cuda/${CUDA_VERSION}/Prod/network_installers/mac/x86_64/cuda_mac_installer_tk.tar.gz
-    fi
-    if ! [ -f cuda_mac_installer_drv.tar.gz ]; then
-        curl -O -# http://developer.download.nvidia.com/compute/cuda/${CUDA_VERSION}/Prod/network_installers/mac/x86_64/cuda_mac_installer_drv.tar.gz
-    fi
-    sudo tar -zxf cuda_mac_installer_tk.tar.gz -C /;
-    sudo tar -zxf cuda_mac_installer_drv.tar.gz -C /;
-    # TODO: Don't delete the tarballs to cache the package, if we can spare the space
-    rm -f cuda_mac_installer_tk.tar.gz cuda_mac_installer_drv.tar.gz
-    # Now head back to work directory
-    cd $TRAVIS_BUILD_DIR
+
+    # Install CUDA
+    # Use solution from https://github.com/JuliaGPU/CUDAapi.jl/pull/81/files
+    declare -A installers
+    installers["7.5"]="http://developer.download.nvidia.com/compute/cuda/7.5/Prod/local_installers/cuda_7.5.27_mac.dmg"
+    installers["8.0"]="https://developer.nvidia.com/compute/cuda/8.0/Prod2/local_installers/cuda_8.0.61_mac-dmg"
+    installers["9.0"]="https://developer.nvidia.com/compute/cuda/9.0/Prod/local_installers/cuda_9.0.176_mac-dmg"
+    installers["9.1"]="https://developer.nvidia.com/compute/cuda/9.1/Prod/local_installers/cuda_9.1.128_mac"
+    installers["9.2"]="https://developer.nvidia.com/compute/cuda/9.2/Prod/local_installers/cuda_9.2.64_mac"
+    installers["10.0"]="https://developer.nvidia.com/compute/cuda/10.0/Prod/local_installers/cuda_10.0.130_mac"
+    installers["10.1"]="https://developer.nvidia.com/compute/cuda/10.1/Prod/local_installers/cuda_10.1.105_mac.dmg"
+    installer=${installers[$CUDA]}
+    wget -O cuda.dmg "$installer"
+
+    brew install p7zip
+    7z x cuda.dmg
+    [[ -f 5.hfs ]] && 7z x 5.hfs
+
+    brew install gnu-tar
+    sudo gtar -x --skip-old-files -f CUDAMacOSXInstaller/CUDAMacOSXInstaller.app/Contents/Resources/payload/cuda_mac_installer_tk.tar.gz -C /
 
     # Install latex.
     export PATH="/usr/texbin:${PATH}:/usr/bin"
